Remove LegacyND2Reader and options
