+++ components/bio-formats-plugins/src/loci/plugins/config/ND2Widgets.java
@@ -35,7 +35,7 @@ import java.awt.event.ItemListener;
 
 import javax.swing.JCheckBox;
 
-import loci.formats.in.NativeND2Reader;
+import loci.formats.in.ND2Reader;
 import loci.plugins.util.LociPrefs;
 
 /**
@@ -53,23 +53,16 @@ public class ND2Widgets implements IFormatWidgets, ItemListener {
   // -- Constructor --
 
   public ND2Widgets() {
-    boolean nikon = Prefs.get(LociPrefs.PREF_ND2_NIKON, false);
-
-    String legacyLabel = "Nikon";
-    JCheckBox legacyBox = new JCheckBox(
-      "Use Nikon's ND2 library instead of native ND2 support", nikon);
-    legacyBox.addItemListener(this);
-
     boolean chunkmap = Prefs.get(LociPrefs.PREF_ND2_CHUNKMAP,
-      NativeND2Reader.USE_CHUNKMAP_DEFAULT);
+      ND2Reader.USE_CHUNKMAP_DEFAULT);
 
     String chunkmapLabel = "Chunkmap";
     JCheckBox chunkmapBox = new JCheckBox(
       "Use chunkmap table to read image offsets", chunkmap);
     chunkmapBox.addItemListener(this);
 
-    labels = new String[] {legacyLabel, chunkmapLabel};
-    widgets = new Component[] {legacyBox, chunkmapBox};
+    labels = new String[] {chunkmapLabel};
+    widgets = new Component[] {chunkmapBox};
   }
 
   // -- IFormatWidgets API methods --
@@ -89,10 +82,7 @@ public class ND2Widgets implements IFormatWidgets, ItemListener {
   @Override
   public void itemStateChanged(ItemEvent e) {
     JCheckBox box = (JCheckBox) e.getSource();
-    if (box.equals(getWidgets()[0])) {
-      Prefs.set(LociPrefs.PREF_ND2_NIKON, box.isSelected());
-    }
-    else if (box.equals(getWidgets()[1])) {
+    if (box.equals(getWidgets()[1])) {
       Prefs.set(LociPrefs.PREF_ND2_CHUNKMAP, box.isSelected());
     }
   }
